# backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: StudyApp API
  version: "1.0.0"
  description: Scan-to-Quiz educational platform API
servers:
  - url: http://localhost:8080
    description: Development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - VALIDATION_ERROR
            - CONFLICT
            - RATE_LIMIT_EXCEEDED
            - FEATURE_NOT_AVAILABLE
            - SCAN_LIMIT_EXCEEDED
            - INTERNAL_ERROR
            - PAYMENT_FAILED
        message:
          type: string
        details:
          nullable: true
        trace_id:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    CheckoutRequest:
      type: object
      required: [planId]
      properties:
        planId:
          type: string
        provider:
          type: string
    CheckoutResponse:
      type: object
      properties:
        orderId:
          type: string
        amount:
          type: integer
          format: int64
        currency:
          type: string
        keyId:
          type: string
        provider:
          type: string
        payuUrl:
          type: string
        formData:
          type: object
          additionalProperties:
            type: string
...
    SelectGoalRequest:
      type: object
      required: [goalId]
      properties:
        goalId:
          type: integer
        targetDate:
          type: string
          format: date
    MyGoal:
      type: object
      properties:
        studentGoalId:
          type: integer
        goalId:
          type: integer
        name:
          type: string
        examName:
          type: string
        targetDate:
          type: string
        status:
          type: string
        progress:
          type: object
...
    SubmitPracticeRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
        submittedAt:
          type: string
          format: date-time
    NotificationPreferences:
      type: object
      properties:
        push_enabled:
          type: boolean
        email_enabled:
          type: boolean
        sms_enabled:
          type: boolean
        max_push_per_day:
          type: integer
        max_email_per_week:
          type: integer
    RegisterDeviceRequest:
      type: object
      required: [token, platform]
      properties:
        token:
          type: string
        platform:
          type: string
        app_version:
          type: string
        os_version:
          type: string
    SendNotificationRequest:
      type: object
      required: [channel, category, template_key]
      properties:
        channel:
          type: string
        category:
          type: string
        template_key:
          type: string
        data:
          type: object
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /auth/devices/register:
    post:
      summary: Register FCM device token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
      responses:
        "201":
          description: Device registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /billing/checkout:
    post:
      summary: Create payment checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
...
  /user/preferences:
    get:
      summary: Get notification preferences
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
    put:
      summary: Update notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferences"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
  /notifications/send:
    post:
      summary: Queue a notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendNotificationRequest"
      responses:
        "202":
          description: Notification queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /goals/select:
    post:
      summary: Select or switch career goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectGoalRequest"
      responses:
        "200":
          description: Goal selected
        "403":
          description: Feature not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goals/my:
    get:
      summary: Get active career goal
      responses:
        "200":
          description: Active goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyGoal"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Abandon active career goal
      responses:
        "200":
          description: Goal abandoned
...
  /goals/my/practice/{setId}/submit:
    post:
      summary: Submit daily practice answers
      parameters:
        - name: setId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitPracticeRequest"
      responses:
        "200":
          description: Practice scored
...